afxdp/cos/queue_ops/pop: relocate inline tests to pop_tests.rs
pop.rs reaches 1,988 LOC with a 1,835-LOC `#[cfg(test)] mod tests` block.
Same playbook as #1080-#1089: extract to a sibling pop_tests.rs.

After: pop.rs drops from 1,988 → 156 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
